requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
unrealspeech
cloudinary
ffmpeg-python
pydub
azure-cognitiveservices-speech
librosa